# packages/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Asset Management Platform API",
    "version": "1.0.0",
    "description": "Domain-oriented API for the Asset Management Platform POC. Three domains: auth (Platform Core), billing (Billing team), accounts (Open Account team)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Exchange credentials for a session. Sets httpOnly access+refresh cookies and a readable csrf cookie; returns the user and csrfToken to echo on state-changing requests.",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Rotate the refresh token and issue a fresh access token + csrf. Requires X-CSRF-Token header matching csrf cookie.",
        "tags": [
          "Auth"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "New session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid refresh token"
          },
          "403": {
            "description": "CSRF token mismatch"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Revoke the current session family and clear session cookies. Requires X-CSRF-Token.",
        "tags": [
          "Auth"
        ],
        "responses": {
          "204": {
            "description": "Logged out"
          },
          "403": {
            "description": "CSRF token mismatch"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Get the current user (validates the access cookie)",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/api/auth/demo-credentials": {
      "get": {
        "summary": "Demo-only (NODE_ENV !== production): list hardcoded credentials for the login page",
        "tags": [
          "Auth"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "List of demo credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DemoCredential"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Disabled in production"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "amp_access_token"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "operator",
          "analyst"
        ]
      },
      "AuthenticatedUser": {
        "type": "object",
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "tenantId"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "tenantId": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "user",
          "csrfToken"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          },
          "csrfToken": {
            "type": "string"
          }
        }
      },
      "DemoCredential": {
        "type": "object",
        "required": [
          "email",
          "password",
          "role"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    }
  }
}